Optimize rasterizer with TriangleEdge wrapper
This wrapped edge class/struct enbles stepping in the actual rasterization so th expensive edge function isn't re calculated entirely for each pixel

This wrapper also sets the stage up for interpolation later

+++ src/Engine/Math/Matrix4.hpp
@@ -1,6 +1,8 @@
 #ifndef MATRIX_HPP
 #define MATRIX_HPP
 
+#include <cstdint>
+
 namespace Engine
 {
 	namespace Math
@@ -19,9 +21,9 @@ namespace Engine
 
 			constexpr Matrix4& operator*=(const Matrix4& other) noexcept
 			{
-				for (int y = 0; y < 4; y++)
+				for (uint8_t y = 0; y < 4; y++)
 				{
-					for (int x = 0; x < 4; x++)
+					for (uint8_t x = 0; x < 4; x++)
 					{
 						values[y][x] =
 							values[y][0] * other.values[0][x] +
